move config handling to cs.py
